Create test for counting the number of cards in the deck
