+++ test/Card-test.js
@@ -3,9 +3,17 @@ const expect = chai.expect;
 
 const { 
   createCard, 
-  evaluateGuess 
+  evaluateGuess
 } = require('../src/card');
-const { subCards } = require('../test/subdata');
+const { 
+  subCards,
+  refCard1,
+  refCard2,
+  refCard3,
+  card1,
+  card2,
+  card3,
+ } = require('../test/subdata');
 
 describe('card', function() {
   it('should be a function', function() {
@@ -54,16 +62,18 @@ describe('turn', function() {
 
 describe('deck', function() {
   it('should create a deck if provided with an array of cards', function() {
-    const refCard1 = subCards[0];
-    const refCard2 = subCards[1];
-    const refCard3 = subCards[2];
-    const card1 = createCard(refCard1.id, refCard1.question, refCard1.answers, refCard1.correctAnswer);
-    const card2 = createCard(refCard2.id, refCard2.question, refCard2.answers, refCard2.correctAnswer);
     const cards = [card1, card2];
     const deck = createDeck(cards);
     expect(deck.card1.id).to.equal(1);
     expect(deck.card1.question).to.equal("What allows you to define a set of related information using key-value pairs?");
     expect(decl.card2.answers).to.equal(["array", "object", "function"])
     expect(deck.card2.correctAnswer).to.equal("mutator method");
+  })
+
+  it('should be able to count the number of cards in the deck', function() {
+    const cards = [card1, card2, card3];
+    const deck = createDeck(cards);
+    const cardCount = countCards(deck);
+    expect(cardCount).to.equal(3);
   })
 })
